% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{addCohortSubsetDefinition}
\alias{addCohortSubsetDefinition}
\title{Add cohort subset definition to a cohort definition set}
\usage{
addCohortSubsetDefinition(cohortDefinitionSet, cohortSubsetDefintion)
}
\arguments{
\item{cohortDefinitionSet}{data.frame that conforms to CohortDefinitionSet}

\item{cohortSubsetDefintion}{CohortSubsetDefinition instance}
}
\description{
Given a subset definition and cohort definition set, this function returns a modified cohortDefinitionSet
That contains cohorts that's have parent's contained within the base cohortDefinitionSet

Also adds the columns subsetParent and isSubset that denote if the cohort is a subset and what the parent definition
is.
}
