% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedCohort.R
\name{addCombinedCohort}
\alias{addCombinedCohort}
\title{Add cohort subset definition to a cohort definition set}
\usage{
addCombinedCohort(
  cohortDefinitionSet,
  combinedCohortDefiniton,
  overwriteExisting = FALSE
)
}
\arguments{
\item{cohortDefinitionSet}{data.frame that conforms to CohortDefinitionSet}

\item{combinedCohortDefiniton}{CombinedCohortDefinition instance}

\item{overwriteExisting}{Overwrite existing subset definition of the same definitionId if present}
}
\description{
Given a subset definition and cohort definition set, this function returns a modified cohortDefinitionSet
That contains cohorts that's have parent's contained within the base cohortDefinitionSet

Also adds the columns subsetParent and isSubset that denote if the cohort is a subset and what the parent definition
is.
}
