% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedCohort.R
\name{CombinedCohortDef}
\alias{CombinedCohortDef}
\title{A CombinedCohortDef}
\description{
Defines the ID, name and operation to produce a new cohort.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortId}}{The list of cohorts to apply in this group.}

\item{\code{cohortName}}{the name given to this outcome definition}

\item{\code{expression}}{The combo operator that is the root of the definition.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CombinedCohortDef-new}{\code{CombinedCohortDef$new()}}
\item \href{#method-CombinedCohortDef-toList}{\code{CombinedCohortDef$toList()}}
\item \href{#method-CombinedCohortDef-toJSON}{\code{CombinedCohortDef$toJSON()}}
\item \href{#method-CombinedCohortDef-isEqualTo}{\code{CombinedCohortDef$isEqualTo()}}
\item \href{#method-CombinedCohortDef-clone}{\code{CombinedCohortDef$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortDef-new"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortDef-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance, using the provided data param if provided.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortDef$new(data = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the data (as a json string or list) to initialize with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortDef-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortDef-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
To JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortDef$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortDef-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortDef-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
Is Equal to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortDef$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortDef-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortDef-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare CombinedCohortDef to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortDef$isEqualTo(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{CombinedCohortDef instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortDef-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortDef-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortDef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
