% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedCohort.R
\name{CombinedCohortOp}
\alias{CombinedCohortOp}
\title{A Combined cohort operation used to UNION or INTERSECT cohorts.  Note: only UNION supported.}
\description{
Defines a UNION or INTERSECT on a set of cohorts.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{targetCohortIds}}{The list of cohorts to apply in this group.}

\item{\code{opType}}{The group operation, either 'union' or 'intersect'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CombinedCohortOp-new}{\code{CombinedCohortOp$new()}}
\item \href{#method-CombinedCohortOp-toList}{\code{CombinedCohortOp$toList()}}
\item \href{#method-CombinedCohortOp-toJSON}{\code{CombinedCohortOp$toJSON()}}
\item \href{#method-CombinedCohortOp-isEqualTo}{\code{CombinedCohortOp$isEqualTo()}}
\item \href{#method-CombinedCohortOp-getDependentCohortIds}{\code{CombinedCohortOp$getDependentCohortIds()}}
\item \href{#method-CombinedCohortOp-clone}{\code{CombinedCohortOp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-new"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance, using the provided data param if provided.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$new(data = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{the data (as a json string or list) to initialize with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
To JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
Is Equal to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare CombinedCohortGroup to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$isEqualTo(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{CombinedCohortGroup instance
getDependentCohortIds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-getDependentCohortIds"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-getDependentCohortIds}{}}}
\subsection{Method \code{getDependentCohortIds()}}{
Gets the dependent cohortIds from this operation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$getDependentCohortIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedCohortOp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedCohortOp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedCohortOp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
