% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceRWrapper.R
\name{createCirceExpressionFromFile}
\alias{createCirceExpressionFromFile}
\title{Create the Circe cohort expression from a JSON file}
\usage{
createCirceExpressionFromFile(filePath)
}
\arguments{
\item{filePath}{The file path containing the Circe JSON file}
}
\description{
This function constructs a Circe cohort expression from a JSON file for use
with other CirceR functions.
}
