% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTables.R
\name{createCohortTables}
\alias{createCohortTables}
\title{Create cohort tables}
\usage{
createCohortTables(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames = getCohortTableNames()
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{The schema to hold the cohort tables. Note that for
SQL Server, this should include both the database and schema
name, for example 'scratch.dbo'.}

\item{cohortTableNames}{The names of the cohort tables. See \code{\link{getCohortTableNames}}
for more details.}
}
\description{
This function creates an empty cohort table and empty tables for
cohort statistics. NOTE: This function will drop the tables if they 
already exist in your cohortDatabaseSchema. You can use 
\code{\link[DatabaseConnector::getTableNames]{DatabaseConnector::getTableNames}}
to check if these tables exist before calling this function.
}
