% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedCohort.R
\name{createCombinedCohortDef}
\alias{createCombinedCohortDef}
\title{Defines a combined cohort using combined cohort operations}
\usage{
createCombinedCohortDef(cohortId, cohortName, expression)
}
\arguments{
\item{cohortId}{The output cohort id from applying the combine expression.}

\item{cohortName}{The output cohort name}

\item{expression}{The combine operator that will yield the final cohort.}
}
\description{
Creates an instance of CombinedCohortDef with the provided cohortId, cohortName and the combine operator expression
}
