% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedCohort.R
\name{createCombinedCohortOp}
\alias{createCombinedCohortOp}
\title{Create CombinedCohortOp instance}
\usage{
createCombinedCohortOp(targetCohortIds, opType)
}
\arguments{
\item{targetCohortIds}{list of target cohort IDs to combine in this operation}

\item{opType}{The op type of this cohort combination, can only be 'union'}
}
\description{
A factory function to create CombinedCohortOp
}
