% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createDemographicSubset}
\alias{createDemographicSubset}
\title{Create Demographic Subset Operator}
\usage{
createDemographicSubset(...)
}
\arguments{
\item{...}{Arguments passed to the underlying operator.}
}
\description{
Subset cohorts using specified limit criteria.
deprecated This function is deprecated. Please use `createDemographicSubsetOperator()` instead.
}
