% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinitionSet.R
\name{createEmptyCohortDefinitionSet}
\alias{createEmptyCohortDefinitionSet}
\title{Create an empty cohort definition set}
\usage{
createEmptyCohortDefinitionSet()
}
\value{
Returns an empty cohort set data.frame
}
\description{
This function creates an empty cohort set data.frame for use
with \code{generateCohortSet}.
}
