% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{createEmptyCohortSet}
\alias{createEmptyCohortSet}
\title{Create an empty cohort set}
\usage{
createEmptyCohortSet()
}
\value{
Returns an empty cohort set data.frame
}
\description{
This function creates an empty cohort set data.frame for use
with \code{instantiateCohortSet}.
}
