% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegativeControlCohorts.R
\name{createEmptyNegativeControlOutcomeCohortSet}
\alias{createEmptyNegativeControlOutcomeCohortSet}
\title{Create an empty negative control outcome cohort set}
\usage{
createEmptyNegativeControlOutcomeCohortSet(verbose = FALSE)
}
\arguments{
\item{verbose}{When TRUE, descriptions of each field in the data.frame are
returned}
}
\value{
Invisibly returns an empty negative control outcome cohort set data.frame
}
\description{
This function creates an empty cohort set data.frame for use
with \code{generateNegativeControlOutcomeCohorts}.
}
