% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateUnionIntersection.R
\name{createIntersectionCohortTemplate}
\alias{createIntersectionCohortTemplate}
\title{Create cohort template to intersect multiple cohorts}
\usage{
createIntersectionCohortTemplate(
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  intersectCohortId
)
}
\arguments{
\item{cohortDatabaseSchema}{Schema where cohort tables are stored.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{A vector of `cohort_definition_id` values for the input cohorts.}

\item{intersectCohortId}{The `cohort_definition_id` for the resulting intersection cohort.}
}
\description{
Create cohort template to intersect multiple cohorts
}
