% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createVisitSubset}
\alias{createVisitSubset}
\title{Create Visit Subset}
\usage{
createVisitSubset(name = NULL, visitConceptIds)
}
\arguments{
\item{name}{Name of operation}

\item{visitConceptIds}{List of visit concept Ids}
}
\description{
Subset cohorts using specified Visit criteria
}
