% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinitionSet.R
\name{.getCohortDefinitionSetSpecification}
\alias{.getCohortDefinitionSetSpecification}
\title{Helper function to return the specification description of a 
cohortDefinitionSet}
\usage{
.getCohortDefinitionSetSpecification()
}
\value{
Returns a data.frame that defines a cohortDefinitionSet
}
\description{
This function reads from the cohortDefinitionSetSpecificationDescription.csv
to return a data.frame that describes the required columns in a 
cohortDefinitionSet
}
