% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{.readCsv}
\alias{.readCsv}
\title{Internal read CSV file when control over column formatting is required.}
\usage{
.readCsv(file)
}
\arguments{
\item{file}{The .csv file to read.}
}
\value{
Returns the file contents invisibly.
}
\description{
This function is internal to the package
}
