% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{.writeCsv}
\alias{.writeCsv}
\title{Internal write.csv file when control over column formatting is required.}
\usage{
.writeCsv(x, file, append = FALSE)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{The .csv file to write.}

\item{append}{When TRUE, append the values of x to an existing file.}
}
\value{
Returns the input x invisibly.
}
\description{
This function is internal to the package since the \code{exportCohortStatsTables}
requires additional control over the column formatting which requires that
we bypass the \code{writeCsv} function since it will automatically convert
from camelCase to snake_case.
}
