% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{Endpoint}
\alias{Endpoint}
\title{to List}
\description{
Representation of an endpoint
}
\details{
Endpoint settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{days}}{Int}

\item{\code{coeff}}{Int}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Endpoint-toList}{\code{Endpoint$toList()}}
\item \href{#method-Endpoint-toJSON}{\code{Endpoint$toJSON()}}
\item \href{#method-Endpoint-isEqualTo}{\code{Endpoint$isEqualTo()}}
\item \href{#method-Endpoint-clone}{\code{Endpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare Endpoint to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$isEqualTo(criteria)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{criteria}}{Endpoint instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Endpoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Endpoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Endpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
