% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortAlgebra.R
\name{eraFyCohort}
\alias{eraFyCohort}
\title{Era-fy cohort(s)}
\usage{
eraFyCohort(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  eraconstructorpad = 0
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{An array of cohort Ids in the cohort table to apply cohort era logic.}

\item{eraconstructorpad}{Optional value to pad cohort era construction logic. Default = 0. i.e. no padding.}
}
\description{
Given a table with cohort_definition_id, subject_id, cohort_start_date,
cohort_end_date execute era logic. This will delete and replace the
original rows with the cohort_definition_id(s). edit privileges
to the cohort table is required.
}
