% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportCohortSubsetStatsTables}
\alias{exportCohortSubsetStatsTables}
\title{Export cohort subset statistics tables to the file system}
\usage{
exportCohortSubsetStatsTables(
  connectionDetails,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames = getCohortTableNames(),
  cohortSubsetStatisticsFolder,
  snakeCaseToCamelCase = TRUE,
  fileNamesInSnakeCase = FALSE,
  databaseId = NULL,
  minCellCount = 5,
  tablePrefix = ""
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTableNames}{The names of the cohort tables. See \code{\link{getCohortTableNames}}
for more details.}

\item{cohortSubsetStatisticsFolder}{The path to the folder where the cohort subset statistics
results will be written.}

\item{snakeCaseToCamelCase}{Should column names in the exported files
convert from snake_case to camelCase? Default is FALSE}

\item{fileNamesInSnakeCase}{Should the exported files use snake_case? Default is FALSE}

\item{databaseId}{Optional - when specified, the databaseId will be added
to the exported results}

\item{minCellCount}{To preserve privacy: the minimum number of subjects contributing
to a count before it can be included in the results. If the
count is below this threshold, it will be set to `-minCellCount`.}

\item{tablePrefix}{Optional - allows to append a prefix to the exported
file names.}
}
\description{
This function retrieves the data from the cohort subset statistics table and
writes them to the subset statistics folder specified in the function call.
}
