% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{generateCohort}
\alias{generateCohort}
\title{Generates a cohort}
\usage{
generateCohort(
  cohortId = NULL,
  cohortSet,
  connection = NULL,
  connectionDetails = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema,
  cohortDatabaseSchema,
  cohortTableNames,
  inclusionStatisticsFolder,
  incremental,
  recordKeepingFile
)
}
\arguments{
\item{cohortId}{The cohortId in the list of \code{cohortSet}}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{inclusionStatisticsFolder}{The folder where the inclusion rule statistics are stored. Can be
left NULL if you do not wish to export the inclusion rule statistics}

\item{incremental}{Create only cohorts that haven't been created before?}

\item{recordKeepingFile}{If \code{incremental = TRUE}, this file will contain
information on cohorts already generated}

\item{cohortTable}{Name of the cohort table.}
}
\description{
This function is used by \code{generateCohortSet} to generate a cohort
against the CDM.
}
