% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{getCohortDefinitionSetFromPackage}
\alias{getCohortDefinitionSetFromPackage}
\title{Get a cohort definition set embedded in a package}
\usage{
getCohortDefinitionSetFromPackage(
  packageName,
  fileName = "settings/CohortsToCreate.csv"
)
}
\value{
Returns a cohort set data.frame
}
\description{
This function supports the legacy way of storing a cohort defnition set in a package,
with a CSV file, JSON files, and SQL files in the `inst` folder.
}
