% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{getCohortInclusionRules}
\alias{getCohortInclusionRules}
\title{Get Cohort Inclusion Rules from a cohort definition set}
\usage{
getCohortInclusionRules(cohortDefinitionSet)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}
}
\description{
This function returns a data frame of the inclusion rules defined
in a cohort definition set.
}
