% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipeFunctions.R
\name{getExclusionSubsetDefinitionIds}
\alias{getExclusionSubsetDefinitionIds}
\title{Get Exclusion Subset Definition Ids}
\usage{
getExclusionSubsetDefinitionIds(cohortDefinitionSet)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}
}
\description{
Get the exlcusion on index subset definition ids from a cohort definition set (if any have been added)
Useful if keeping track in a script with complex business logic around what a cohort definition is for
}
