% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{getRequiredTasks}
\alias{getRequiredTasks}
\title{Get a list of tasks required when running in incremental mode}
\usage{
getRequiredTasks(..., checksum, recordKeepingFile)
}
\arguments{
\item{...}{Parameter values used to identify the key 
in the incremental record keeping file}

\item{checksum}{The checksum representing the operation to check}

\item{recordKeepingFile}{A file path to a CSV file containing the record
keeping information.}
}
\value{
Returns a list of outstanding tasks based on inspecting the full contents
of the record keeping file
}
\description{
This function will attempt to check the \code{recordKeepingFile}
to determine if a list of operations have completed by comparing the 
keys passed into the function with the checksum supplied
}
