% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetDefinitions.R
\name{getSubsetDefinitions}
\alias{getSubsetDefinitions}
\title{Get cohort subset definitions from a cohort definition set}
\usage{
getSubsetDefinitions(cohortDefinitionSet)
}
\arguments{
\item{cohortDefinitionSet}{A valid cohortDefinitionSet}
}
\value{
list of cohort subset definitions or empty list
}
\description{
Get the subset definitions (if any) applied to a cohort definition set.
Note that these subset definitions are a copy of those applied to the cohort set.
Modifying these definitions will not modify the base cohort set.
To apply a modification, reapply the subset definition to the cohort definition set data.frame with
\code{addCohortSubsetDefinition} with `overwriteExisting = TRUE`.
}
