% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{instantiateCohortSet}
\alias{instantiateCohortSet}
\title{Instantiate a set of cohorts}
\usage{
instantiateCohortSet(
  connectionDetails = NULL,
  connection = NULL,
  numThreads = 1,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortSet = NULL,
  inclusionStatisticsFolder = NULL,
  createCohortTable = FALSE,
  incremental = FALSE,
  incrementalFolder = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{numThreads}{Specify the number of threads for cohort generation. Currently
this only supports single threaded operations.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortSet}{The \code{cohortSet} argument must be a data frame with 
the following columns: \describe{\item{cohortId}{The 
unique integer identifier of the cohort} 
\item{cohortFullName}{The cohort's full name}
\item{sql}{The OHDSI-SQL used to instantiate the cohort}
\item{json}{The json column must represent a Circe cohort 
definition. This field is only required when you would 
like to generate a cohort that includes inclusion statistics
since the names of the inclusion rules are parsed from 
this JSON property.} }}

\item{inclusionStatisticsFolder}{The folder where the inclusion rule statistics are stored. Can be
left NULL if you do not wish to export the inclusion rule statistics}

\item{createCohortTable}{Create the cohort table? If \code{incremental = TRUE} and the
table already exists this will be skipped.}

\item{incremental}{Create only cohorts that haven't been created before?}

\item{incrementalFolder}{If \code{incremental = TRUE}, specify a folder where records are
kept of which definition has been executed.}
}
\description{
This function instantiates a set of cohorts in the cohort table and where
specified the inclusion rule statistics are computed and stored in the
\code{inclusionStatisticsFolder}.
}
