% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{isCamelCase}
\alias{isCamelCase}
\title{Used to check if a string is in lower camel case}
\usage{
isCamelCase(x)
}
\arguments{
\item{x}{The string to evaluate}
}
\value{
TRUE if the string is in lower camel case
}
\description{
This function is used check if a string conforms to
the lower camel case format.
}
