% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinitionSet.R
\name{isCohortDefinitionSet}
\alias{isCohortDefinitionSet}
\title{Is the data.frame a cohort definition set?}
\usage{
isCohortDefinitionSet(x)
}
\value{
Returns TRUE if the input is a cohortDefinitionSet
}
\description{
This function checks a data.frame to verify it holds the expected format
for a cohortDefinitionSet.
}
