% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{isFormattedForDatabaseUpload}
\alias{isFormattedForDatabaseUpload}
\title{Used to check a data.frame or tibble to see if it conforms to 
the standard for uploading to a database}
\usage{
isFormattedForDatabaseUpload(x)
}
\arguments{
\item{x}{A data frame or tibble}
}
\value{
Returns TRUE if all columns are lower snake case otherwise returns
FALSE and warns on the column names that violate the rule.
}
\description{
This function is used to check a data.frame or tibble to ensure all 
column names are lower snake case.
}
