% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{isSnakeCase}
\alias{isSnakeCase}
\title{Used to check if a string is in lower snake case}
\usage{
isSnakeCase(x)
}
\arguments{
\item{x}{The string to evaluate}
}
\value{
TRUE if the string is in lower snake case
}
\description{
This function is used check file and field names
to ensure the proper casing is in use.
}
