% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{isTaskRequired}
\alias{isTaskRequired}
\title{Is a task required when running in incremental mode}
\usage{
isTaskRequired(..., checksum, recordKeepingFile, verbose = TRUE)
}
\arguments{
\item{...}{Parameter values used to identify the key 
in the incremental record keeping file}

\item{checksum}{The checksum representing the operation to check}

\item{recordKeepingFile}{A file path to a CSV file containing the record
keeping information.}

\item{verbose}{When TRUE, this function will output if a particular operation
has completed based on inspecting the recordKeepingFile.}
}
\value{
Returns TRUE if the operation has completed according to the contents of
the record keeping file.
}
\description{
This function will attempt to check the \code{recordKeepingFile}
to determine if an individual operation has completed by comparing the 
keys passed into the function with the checksum supplied
}
