% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{migrateDataModel}
\alias{migrateDataModel}
\title{Migrate Data model}
\usage{
migrateDataModel(connectionDetails, databaseSchema, tablePrefix = "")
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{databaseSchema}{String schema where database schema lives}

\item{tablePrefix}{(Optional) Use if a table prefix is used before table names (e.g. "cg_")}
}
\description{
Migrate data from current state to next state

It is strongly advised that you have a backup of all data (either sqlite files, a backup database (in the case you
are using a PostgreSQL backend) or have kept the csv/zip files from your data generation.
}
