% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopCdmPerson}
\alias{omopCdmPerson}
\title{OMOP CDM Person Sample Data}
\format{
A data frame with 12 rows and 5 variables:
\describe{
  \item{person_id}{A unique identifier for the person}
  \item{gender_concept_id}{An integer concept ID representing the person's gender}
  \item{year_of_birth}{Year of birth}
  \item{race_concept_id}{An integer concept ID representing the person's race}
  \item{ethnicity_concept_id}{An integer concept ID representing the person's ethnicity}
}
}
\source{
Fictional data for demonstration.
}
\usage{
omopCdmPerson
}
\description{
A data set containing sample persons
}
\keyword{datasets}
