% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceRWrapper.R
\name{readCirceExpressionJsonFile}
\alias{readCirceExpressionJsonFile}
\title{Read a JSON file for use with CirceR}
\usage{
readCirceExpressionJsonFile(filePath)
}
\arguments{
\item{filePath}{The file path containing the Circe JSON file}
}
\description{
This function wraps the default readlines call for calling CirceR.
}
