% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{recordTasksDone}
\alias{recordTasksDone}
\title{Record a task as complete}
\usage{
recordTasksDone(..., checksum, recordKeepingFile, incremental = TRUE)
}
\arguments{
\item{...}{Parameter values used to identify the key 
in the incremental record keeping file}

\item{checksum}{The checksum representing the operation to check}

\item{recordKeepingFile}{A file path to a CSV file containing the record
keeping information.}

\item{incremental}{When TRUE, this function will record tasks otherwise
it will return without attempting to perform any action}
}
\description{
This function will record a task as completed in the \code{recordKeepingFile}
}
