% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{saveIncremental}
\alias{saveIncremental}
\title{Used in incremental mode to save values to a file}
\usage{
saveIncremental(data, fileName, ...)
}
\arguments{
\item{data}{The data to record in the file}

\item{fileName}{A CSV holding results in the same structure as the data
parameter}

\item{...}{Parameter values used to identify the key 
in the results file}
}
\description{
When running in incremental mode, we may need to update results in a CSV
file. This function will replace the \code{data} in \code{fileName} based
on the key parameters
}
