% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortConstruction.R
\name{sqlContainsInclusionRuleStats}
\alias{sqlContainsInclusionRuleStats}
\title{Detects if the SQL indicate to compute inclusion rule statistics}
\usage{
sqlContainsInclusionRuleStats(sql)
}
\arguments{
\item{sql}{A string containing the SQL used to instantiate a cohort. This
code assumes that the SQL has not been rendered using SqlRender
in order to detect tokens that indicate the generation of
inclusion rule statistics in addition to the cohort.}
}
\description{
This function takes as a parameter a SQL script used to generate a cohort
and performs a string search for tokens that indicate to generate the inclusion 
statistics. This SQL is usually generated by circe-be.

This function also assumes that the SQL passed into the function has not been
translated to a specific SQL dialect.
}
