% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortAlgebra.R
\name{unionCohort}
\alias{unionCohort}
\title{union cohort}
\usage{
unionCohort(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds,
  newCohortId
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort table resides. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{An array of one or more cohort ids}

\item{newCohortId}{The cohort id of the new derived cohort. This should not already
exist in the cohort table.}
}
\description{
Given cohortIds of instantiated cohorts in a cohort table,
create a new derived cohort in the same table with a new cohort_definition_id
that is derived by performing a union of the given cohortIds and ensuring
that they conform to cohort era rules by applying era-fy logic.
}
