% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{Window}
\alias{Window}
\title{to List}
\description{
Representation of a time window
Based on https://github.com/OHDSI/circe-be/blob/master/src/main/java/org/ohdsi/circe/cohortdefinition/Window.java
}
\details{
Window settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{start}}{Endpoint}

\item{\code{end}}{Endpoint}

\item{\code{useIndexEnd}}{Boolean}

\item{\code{useEventEnd}}{Boolean}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Window-toList}{\code{Window$toList()}}
\item \href{#method-Window-toJSON}{\code{Window$toJSON()}}
\item \href{#method-Window-isEqualTo}{\code{Window$isEqualTo()}}
\item \href{#method-Window-clone}{\code{Window$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Window-toList"></a>}}
\if{latex}{\out{\hypertarget{method-Window-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Window$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Window-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Window-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Window$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Window-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-Window-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare Window to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Window$isEqualTo(criteria)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{criteria}}{Window instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Window-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Window-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Window$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
