% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{writeCsv}
\alias{writeCsv}
\title{Used to write a csv file}
\usage{
writeCsv(x, file)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{The csv file to write.}
}
\value{
Returns the input x invisibly.
}
\description{
This function is used to centralize the function for writing
CSV files across the HADES ecosystem.
}
