% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateCohorts.R
\name{addCohortTemplateDefintion}
\alias{addCohortTemplateDefintion}
\title{Add Cohort template definition to cohort set}
\usage{
addCohortTemplateDefintion(
  cohortDefinitionSet = createEmptyCohortDefinitionSet(),
  cohortTemplateDefintion
)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}

\item{cohortTemplateDefintion}{An instance of CohortTemplateDefinition (or subclass).
See [@seealso [createCohortTemplateDefinition()]].}
}
\description{
Adds a cohort template definition to an existing cohort definition set or creates one if none provided
}
\seealso{
Other templateCohorts: 
\code{\link{CohortTemplateDefinition}},
\code{\link{addSqlCohortDefinition}()},
\code{\link{createCohortTemplateDefintion}()},
\code{\link{getTemplateDefinitions}()}
}
\concept{templateCohorts}
