% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipeFunctions.R
\name{addExcludeOnIndexSubsetDefinition}
\alias{addExcludeOnIndexSubsetDefinition}
\title{Add exclude on index subset definition}
\usage{
addExcludeOnIndexSubsetDefinition(
  cohortDefinitionSet,
  name,
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  targetCohortIds,
  exclusionCohortIds,
  exclusionWindow = 0,
  definitionId,
  cohortCombinationOperator = "any"
)
}
\arguments{
\item{cohortDefinitionSet}{Data frame with columns: cohortId, cohortName, sql, and optionally json.}

\item{exclusionCohortIds}{cohort ids to exlcude members of target from}

\item{exclusionWindow}{Days Default is 0 (target index date). by changing this
you can adjust the period around target index for which you would
exclude members.}

\item{cohortCombinationOperator}{if more than one cohort is used, combine them all with any or only
exclude if they are in a single cohort definition}
}
\description{
The purpose of this subset recipie is to exclude all individuals if their index aligns with the specified
exclusion cohort ids.This may be used in situations where an outcome cohort may contain individuals treated for a target medication,
complicating calculation of incidence rates.
}
