% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipeFunctions.R
\name{addExcludeOnIndexSubsetDefinition}
\alias{addExcludeOnIndexSubsetDefinition}
\title{Add exclude on index subset definition}
\usage{
addExcludeOnIndexSubsetDefinition(
  cohortDefinitionSet,
  name,
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  targetCohortIds,
  exclusionCohortIds,
  exclusionWindow = 0,
  definitionId,
  cohortCombinationOperator = "any"
)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}

\item{name}{name of the subset definition (used in resulting cohort definitions)}

\item{subsetCohortNameTemplate}{template string format for naming resulting cohorts}

\item{targetCohortIds}{Set of integer cohort IDs. Must be within the cohort definition set.}

\item{exclusionCohortIds}{cohort ids to exlcude members of target from}

\item{exclusionWindow}{Days Default is 0 (target index date). by changing this
you can adjust the period around target index for which you would
exclude members.}

\item{definitionId}{Unique integer Id of the subset definition}

\item{cohortCombinationOperator}{Logic for multiple indication cohort IDs: any (default) or all.}
}
\description{
The purpose of this subset recipie is to exclude all individuals if their index aligns with the specified
exclusion cohort ids.This may be used in situations where an outcome cohort may contain individuals treated for a target medication,
complicating calculation of incidence rates.
}
