% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipes.R
\name{addIndicationSubsetDefinition}
\alias{addIndicationSubsetDefinition}
\title{Add Indication Subset Definition}
\usage{
addIndicationSubsetDefinition(
  cohortDefinitionSet,
  targetCohortIds,
  indicationCohortIds,
  definitionId,
  name = "study population + first",
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  cohortCombinationOperator = "any",
  lookbackWindowStart = -99999,
  loockbackWindowEnd = 0,
  lookForwardWindowStart = 0,
  lookForwardWindowEnd = 99999,
  genderConceptIds = NULL,
  ageMin = NULL,
  ageMax = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredPriorObservationTime = 365,
  requiredFollowUpTime = 1
)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}

\item{indicationCohortIds}{set of intger cohort ids. Must be within the cohort definition set}

\item{cohortCombinationOperator}{If there is more than one cohort id as an indication condition, what logic should
be required for inclusion - default is any. Alternatively, all prior diseases
can be required in the specified window.
Specifying 'all' can be useful in the case of severe conditions (e.g. individuals with
prior disease that also require prior exposure to a previous drug for which the
targetExposureId is a second line treatment)}

\item{lookbackWindowStart}{start of the lookback period for valid inclusion within. How long prior can
an individual have a condition}

\item{lookForwardWindowStart}{When can the indicated condition end relative to the target Default - day 0}

\item{lookForwardWindowEnd}{When can the indicated condition end relative to the target Default 9999}

\item{ageMin}{Does this population require a minimum age (at target index)?}

\item{ageMax}{Does this population require a maximum age (at target index)?}

\item{studyStartDate}{Exclude patients index prior to this date. Must be in date format - "%Y%m%d" (e.g. 2001/12/25)}

\item{studyEndDate}{Exclude patients with index after this date. Must be in date format - "%Y%m%d" (e.g. 2001/12/25)}

\item{requiredPriorObservationTime}{Observation time in data source required prior to index (a.k.a washout perioid)
Deafults to 365 days.}

\item{requiredFollowUpTime}{Observation time after target index required (typically at least 1 day, default)}

\item{subsetDefinitionId}{The ID if the resulting subset. Note, this must be uniquely applied to the cohort
definition set.}

\item{lookbackWindowEnd}{end of the lookback period for valid inclusion within. This is normally day 0,
index. However, there may be situations where there must be a gap between diagnosis
and indication start.}
}
\description{
Utility pattern for creating cohort subset definitions as a standaridzed approach for indicated drugs.
The general approach is to apply this definition to an exposure or set of exposures such that an individual must
show prior history of some condition prior to recieving treatment.

This is a function designed to make parameterization of study execution clear.

Also attaches an attribute to the cohort definition set
}
