% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipeFunctions.R
\name{addIndicationSubsetDefinition}
\alias{addIndicationSubsetDefinition}
\title{Add Indication Subset Definition}
\usage{
addIndicationSubsetDefinition(
  cohortDefinitionSet,
  targetCohortIds,
  indicationCohortIds,
  definitionId,
  name = "study population + first",
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  cohortCombinationOperator = "any",
  lookbackWindowStart = -99999,
  loockbackWindowEnd = 0,
  lookForwardWindowStart = 0,
  lookForwardWindowEnd = 99999,
  genderConceptIds = NULL,
  ageMin = NULL,
  ageMax = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredPriorObservationTime = 365,
  requiredFollowUpTime = 1
)
}
\arguments{
\item{cohortDefinitionSet}{Data frame with columns: cohortId, cohortName, sql, and optionally json.}

\item{indicationCohortIds}{Set of integer cohort IDs. Must be within the cohort definition set.}

\item{cohortCombinationOperator}{Logic for multiple indication cohort IDs: any (default) or all.}

\item{lookbackWindowStart}{Start of lookback period.}

\item{lookForwardWindowStart}{When the indication can end relative to index; default is 0.}

\item{lookForwardWindowEnd}{When the indication can end relative to index; default is 9999.}

\item{ageMin}{Minimum age at target index.}

\item{ageMax}{Maximum age at target index.}

\item{studyStartDate}{Exclude patients with index prior to this date (format "\%Y\%m\%d").}

\item{studyEndDate}{Exclude patients with index after this date (format "\%Y\%m\%d").}

\item{requiredPriorObservationTime}{Observation time prior to index; default 365.}

\item{requiredFollowUpTime}{Observation time after index; default 1.}

\item{lookbackWindowEnd}{End of lookback period.}

\item{subsetDefinitionId}{Unique ID for the subset}
}
\description{
Utility pattern for creating cohort subset definitions, following a standard approach for indicated drugs.
The approach applies this definition to an exposure or set of exposures, requiring individuals to have
a prior history of a condition before receiving treatment.
This function aims to make parameterization of study execution explicit.
Additionally, it attaches an attribute to the cohort definition set.
}
