% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipeFunctions.R
\name{addIndicationSubsetDefinition}
\alias{addIndicationSubsetDefinition}
\title{Add Indication Subset Definition}
\usage{
addIndicationSubsetDefinition(
  cohortDefinitionSet,
  targetCohortIds,
  indicationCohortIds,
  definitionId,
  name = "study population + first",
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  cohortCombinationOperator = "any",
  lookbackWindowStart = -99999,
  lookbackWindowEnd = 0,
  lookForwardWindowStart = 0,
  lookForwardWindowEnd = 99999,
  genderConceptIds = NULL,
  ageMin = NULL,
  ageMax = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredPriorObservationTime = 365,
  requiredFollowUpTime = 1
)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}

\item{targetCohortIds}{Set of integer cohort IDs. Must be within the cohort definition set.}

\item{indicationCohortIds}{Set of integer cohort IDs. Must be within the cohort definition set.}

\item{definitionId}{Unique integer Id of the subset definition}

\item{name}{name of the subset definition (used in resulting cohort definitions)}

\item{subsetCohortNameTemplate}{template string format for naming resulting cohorts}

\item{cohortCombinationOperator}{Logic for multiple indication cohort IDs: any (default) or all.}

\item{lookbackWindowStart}{Start of lookback period.}

\item{lookbackWindowEnd}{End of lookback period.}

\item{lookForwardWindowStart}{When the indication can end relative to index; default is 0.}

\item{lookForwardWindowEnd}{When the indication can end relative to index; default is 9999.}

\item{genderConceptIds}{Gender concepts to require}

\item{ageMin}{Minimum age at target index.}

\item{ageMax}{Maximum age at target index.}

\item{studyStartDate}{Exclude patients with index prior to this date (format "\%Y\%m\%d").}

\item{studyEndDate}{Exclude patients with index after this date (format "\%Y\%m\%d").}

\item{requiredPriorObservationTime}{Observation time prior to index; default 365.}

\item{requiredFollowUpTime}{Observation time after index; default 1.}
}
\description{
Utility pattern for creating cohort subset definitions, following a standard approach for indicated drugs.
The approach applies this definition to an exposure or set of exposures, requiring individuals to have
a prior history of a condition before receiving treatment.
This function aims to make parameterization of study execution explicit.
Additionally, it attaches an attribute to the cohort definition set.
}
