% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipes.R
\name{addRestrictionSubsetDefinition}
\alias{addRestrictionSubsetDefinition}
\title{Add Restriction Subset Definition}
\usage{
addRestrictionSubsetDefinition(
  cohortDefinitionSet,
  targetCohortIds,
  definitionId,
  name = "study population + first",
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  genderConceptIds = NULL,
  ageMin = NULL,
  ageMax = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredPriorObservationTime = 365,
  requiredFollowUpTime = 1
)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}

\item{ageMin}{Does this population require a minimum age (at target index)?}

\item{ageMax}{Does this population require a maximum age (at target index)?}

\item{studyStartDate}{Exclude patients index prior to this date. Must be in date format - "%Y%m%d" (e.g. 2001/12/25)}

\item{studyEndDate}{Exclude patients with index after this date. Must be in date format - "%Y%m%d" (e.g. 2001/12/25)}

\item{requiredPriorObservationTime}{Observation time in data source required prior to index (a.k.a washout perioid)
Deafults to 365 days.}

\item{requiredFollowUpTime}{Observation time after target index required (typically at least 1 day, default)}
}
\description{
Utility pattern for creating cohort subset definitions as a standaridzed approach for indicated drugs.
Restriction subset definitions are twins of indication definitions. They should apply the same core properites
to a base exposure cohort (i.e. study dates,  required prior observation time, ages, gender) as indications but,
crucially, they do not require history of any prior conidtion(s).

This is useful in the context of comparing drug exposure + indication population, to population as a whole.

The prefered use of this function is to create this in conjunction with the target population.
}
