% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRecipes.R
\name{addRestrictionSubsetDefinition}
\alias{addRestrictionSubsetDefinition}
\title{Add Restriction Subset Definition}
\usage{
addRestrictionSubsetDefinition(
  cohortDefinitionSet,
  targetCohortIds,
  definitionId,
  name = "study population + first",
  subsetCohortNameTemplate = "@baseCohortName - @subsetDefinitionName",
  genderConceptIds = NULL,
  ageMin = NULL,
  ageMax = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  requiredPriorObservationTime = 365,
  requiredFollowUpTime = 1
)
}
\arguments{
\item{cohortDefinitionSet}{Data frame with columns: cohortId, cohortName, sql, and optionally json.}

\item{ageMin}{Minimum age at target index.}

\item{ageMax}{Maximum age at target index.}

\item{studyStartDate}{Exclude patients with index prior to this date (format "%Y%m%d").}

\item{studyEndDate}{Exclude patients with index after this date (format "%Y%m%d").}

\item{requiredPriorObservationTime}{Observation time prior to index; default 365.}

\item{requiredFollowUpTime}{Observation time after index; default 1.}
}
\description{
Utility pattern for creating cohort subset definitions as a standard approach for indicated drugs.
Restriction subset definitions are twins of indication definitions. They should apply the same core properites
to a base exposure cohort (i.e. study dates,  required prior observation time, ages, gender) as indications but,
crucially, they do not require history of any prior condition(s).

This is useful in the context of comparing drug exposure + indication population, to population as a whole.

The prefered use of this function is to create this in conjunction with the target population.
}
