% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateUnion.R
\name{addUnionCohortDefinition}
\alias{addUnionCohortDefinition}
\title{Add union cohort definition to cohort definition set}
\usage{
addUnionCohortDefinition(
  cohortDefinitionSet,
  cohortIds,
  cohortName,
  unionCohortId
)
}
\arguments{
\item{cohortDefinitionSet}{cohort definition set}

\item{cohortIds}{A vector of `cohort_definition_id` values for the input cohorts.}

\item{cohortName}{The Name of the resulting cohort}

\item{unionCohortId}{The `cohort_definition_id` for the resulting union cohort.}
}
\description{
This utility function adds the union of any two or more cohort ids to the cohort definition set with a new id and name.

If a name parameter is not provideded this will be auto generated as the union of the provided cohort id
}
