% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\docType{package}
\name{CohortGenerator-package}
\alias{CohortGenerator}
\alias{CohortGenerator-package}
\title{CohortGenerator: Used to generate OHDSI cohorts}
\description{
Package to generate cohorts and other functions related to cohort generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anthonysena/CohortGenerator}
  \item Report bugs at \url{https://github.com/anthonysena/CohortGenerator/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{asena5@its.jnj.com}

Authors:
\itemize{
  \item Gowtham Rao \email{gowthamrao@gmail.com}
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
