% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\docType{package}
\name{CohortGenerator-package}
\alias{CohortGenerator}
\alias{CohortGenerator-package}
\title{CohortGenerator: Cohort Generation for the OMOP Common Data Model}
\description{
Generate cohorts and subsets using an Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) Database. Cohorts are defined using 'CIRCE' (\url{https://github.com/ohdsi/circe-be}) or SQL compatible with 'SqlRender' (\url{https://github.com/OHDSI/SqlRender}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortGenerator/}
  \item \url{https://github.com/OHDSI/CohortGenerator}
  \item Report bugs at \url{https://github.com/OHDSI/CohortGenerator/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{asena5@its.jnj.com}

Authors:
\itemize{
  \item Jamie Gilbert
  \item Gowtham Rao \email{gowthamrao@gmail.com}
  \item Freddy Avila Cruz
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
