% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortGenerator.R
\docType{package}
\name{CohortGenerator-package}
\alias{CohortGenerator}
\alias{CohortGenerator-package}
\title{CohortGenerator: An R package for cohort generation against the OMOP CDM}
\description{
An R package for that encapsulates the functions for generating cohorts against the OMOP CDM. This package allow for the use of SQL generated using CIRCE-compliant JSON or custom SQL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortGenerator/}
  \item \url{https://github.com/OHDSI/CohortGenerator}
  \item Report bugs at \url{https://github.com/anthonysena/CohortGenerator/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{asena5@its.jnj.com}

Authors:
\itemize{
  \item Gowtham Rao \email{gowthamrao@gmail.com}
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
