% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{CohortSubsetDefinition}
\alias{CohortSubsetDefinition}
\title{Cohort Subset Definition}
\description{
Set of subset definitions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{targetOutcomePairs}}{list of pairs of intgers - (targetCohortId, outputCohortId)}

\item{\code{subsets}}{list of subset operations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortSubsetDefinition-new}{\code{CohortSubsetDefinition$new()}}
\item \href{#method-CohortSubsetDefinition-toList}{\code{CohortSubsetDefinition$toList()}}
\item \href{#method-CohortSubsetDefinition-toJSON}{\code{CohortSubsetDefinition$toJSON()}}
\item \href{#method-CohortSubsetDefinition-addSubsetOperator}{\code{CohortSubsetDefinition$addSubsetOperator()}}
\item \href{#method-CohortSubsetDefinition-getSubsetOperatorById}{\code{CohortSubsetDefinition$getSubsetOperatorById()}}
\item \href{#method-CohortSubsetDefinition-clone}{\code{CohortSubsetDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$new(definition = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{definition}}{json or list representation of object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-addSubsetOperator"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-addSubsetOperator}{}}}
\subsection{Method \code{addSubsetOperator()}}{
add subset to class - checks if equivalent id is present
Will throw an error if a matching ID is found but reference object is different
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$addSubsetOperator(subsetOperator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subsetOperator}}{a SubsetOperator isntance}

\item{\code{overwrite}}{if a subset operator of the same ID is present, replace it with a new definition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-getSubsetOperatorById"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-getSubsetOperatorById}{}}}
\subsection{Method \code{getSubsetOperatorById()}}{
get a subset operator by its id field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$getSubsetOperatorById(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Integer subset id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetDefinition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetDefinition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
