% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{CohortSubsetDefinition}
\alias{CohortSubsetDefinition}
\title{Cohort Subset Definition}
\description{
Set of subset definitions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{outputCohortId}}{integer for output cohort id for subset}

\item{\code{targetCohortIds}}{Integer ofor optional subset of cohort ids to use}

\item{\code{subsets}}{list of subset operations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CohortSubsetDefinition$new()}}
\item \href{#method-toList}{\code{CohortSubsetDefinition$toList()}}
\item \href{#method-toJSON}{\code{CohortSubsetDefinition$toJSON()}}
\item \href{#method-addSubsetOperator}{\code{CohortSubsetDefinition$addSubsetOperator()}}
\item \href{#method-clone}{\code{CohortSubsetDefinition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$new(definition = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{definition}}{json or list representation of object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSubsetOperator"></a>}}
\if{latex}{\out{\hypertarget{method-addSubsetOperator}{}}}
\subsection{Method \code{addSubsetOperator()}}{
add subset to class - checks if equivalent id is present
Will throw an error if a matching ID is found but reference object is different
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$addSubsetOperator(subsetOperator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subsetOperator}}{a SubsetOperator isntance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
