% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{CohortSubsetOperator}
\alias{CohortSubsetOperator}
\title{Cohort Subset Operator}
\description{
A subset of type cohort - subset a population to only those contained within defined cohort
}
\section{Super class}{
\code{\link[CohortGenerator:SubsetOperator]{CohortGenerator::SubsetOperator}} -> \code{CohortSubsetOperator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortIds}}{Integer ids of cohorts to subset to}

\item{\code{cohortCombinationOperator}}{How to combine the cohorts}

\item{\code{negate}}{Inverse the subset rule? TRUE will take the patients NOT in the subset}

\item{\code{windows}}{list of time windows to use when evaluating the subset
cohort relative to the target cohort}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortSubsetOperator-new}{\code{CohortSubsetOperator$new()}}
\item \href{#method-CohortSubsetOperator-toList}{\code{CohortSubsetOperator$toList()}}
\item \href{#method-CohortSubsetOperator-getAutoGeneratedName}{\code{CohortSubsetOperator$getAutoGeneratedName()}}
\item \href{#method-CohortSubsetOperator-clone}{\code{CohortSubsetOperator$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="classname"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-classname'><code>CohortGenerator::SubsetOperator$classname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="getQueryBuilder"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-getQueryBuilder'><code>CohortGenerator::SubsetOperator$getQueryBuilder()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="isEqualTo"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-isEqualTo'><code>CohortGenerator::SubsetOperator$isEqualTo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="print"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-print'><code>CohortGenerator::SubsetOperator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="publicFields"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-publicFields'><code>CohortGenerator::SubsetOperator$publicFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="toJSON"><a href='../../CohortGenerator/html/SubsetOperator.html#method-SubsetOperator-toJSON'><code>CohortGenerator::SubsetOperator$toJSON()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetOperator-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetOperator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$new(definition = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{definition}}{json character or list - definition of subset operator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
instance of object
to List
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetOperator-toList"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetOperator-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
Get auto generated name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetOperator-getAutoGeneratedName"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetOperator-getAutoGeneratedName}{}}}
\subsection{Method \code{getAutoGeneratedName()}}{
name generated from subset operation properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$getAutoGeneratedName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSubsetOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSubsetOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
