% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{CohortSubsetOperator}
\alias{CohortSubsetOperator}
\title{Cohort Subset}
\description{
A subset of type cohort - subset a population to only those contained within defined cohort
}
\section{Super class}{
\code{\link[CohortGenerator:SubsetOperator]{CohortGenerator::SubsetOperator}} -> \code{CohortSubsetOperator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortIds}}{Integer ids of cohorts to subset to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-publicFields}{\code{CohortSubsetOperator$publicFields()}}
\item \href{#method-toList}{\code{CohortSubsetOperator$toList()}}
\item \href{#method-clone}{\code{CohortSubsetOperator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="classname">}\href{../../CohortGenerator/html/SubsetOperator.html#method-classname}{\code{CohortGenerator::SubsetOperator$classname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="initialize">}\href{../../CohortGenerator/html/SubsetOperator.html#method-initialize}{\code{CohortGenerator::SubsetOperator$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="isEqualTo">}\href{../../CohortGenerator/html/SubsetOperator.html#method-isEqualTo}{\code{CohortGenerator::SubsetOperator$isEqualTo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="toJSON">}\href{../../CohortGenerator/html/SubsetOperator.html#method-toJSON}{\code{CohortGenerator::SubsetOperator$toJSON()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-publicFields"></a>}}
\if{latex}{\out{\hypertarget{method-publicFields}{}}}
\subsection{Method \code{publicFields()}}{
publicly settable fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$publicFields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSubsetOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
