% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incremental.R
\name{computeChecksum}
\alias{computeChecksum}
\title{Computes the checksum for a value}
\usage{
computeChecksum(val)
}
\arguments{
\item{val}{The value to hash. It is converted to a character to perform
the hash.}
}
\value{
Returns a string containing the checksum
}
\description{
This is used as part of the incremental operations to hash a value
to store in a record keeping file. This function leverages the md5
hash from the digest package
}
