% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createCohortSubset}
\alias{createCohortSubset}
\title{Create Cohort Subset Operator}
\usage{
createCohortSubset(...)
}
\arguments{
\item{...}{Arguments passed to the underlying operator.}
}
\description{
Subset cohorts using specified limit criteria.
deprecated This function is deprecated. Please use `createCohortSubsetOperator()` instead.
}
