% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createCohortSubset}
\alias{createCohortSubset}
\title{A definition of subset functions to be applied to a set of cohorts}
\usage{
createCohortSubset(id, name, cohortIds)
}
\arguments{
\item{id}{unique integer identifier}

\item{name}{name of operator}

\item{cohortIds}{integer - set of cohort ids to subset to}
}
\value{
a CohortSubset instance
}
\description{
A definition of subset functions to be applied to a set of cohorts
}
