% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createCohortSubsetDefinition}
\alias{createCohortSubsetDefinition}
\title{Create Subset Definition}
\usage{
createCohortSubsetDefinition(name, definitionId, targetOutcomePairs, subsets)
}
\arguments{
\item{name}{Name of definition}

\item{definitionId}{Definition identifier}

\item{targetOutcomePairs}{Vector of pairs targetCohortId, outcomeCohortId}

\item{subsets}{vector of subset instances to apply}
}
\description{
Create subset definition from subset objects
}
\examples{
mySubsets <- c(
 createCohortSubset(id = 12, name = "foo", cohortJson = cohortDefinition)
)

subsetDef <- createSubsetDefinition(c(1,3,4), 11, mySubsets)
}
