% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetDefinitions.R
\name{createCohortSubsetDefinition}
\alias{createCohortSubsetDefinition}
\title{Create Subset Definition}
\usage{
createCohortSubsetDefinition(
  name,
  definitionId,
  subsetOperators,
  identifierExpression = NULL
)
}
\arguments{
\item{name}{Name of definition}

\item{definitionId}{Definition identifier}

\item{subsetOperators}{list of subsetOperator instances to apply}

\item{identifierExpression}{Expression (or string that converts to expression) that returns an id for an output cohort
the default is dplyr::expr(targetId * 1000 + definitionId)}
}
\description{
Create subset definition from subset objects
}
