% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetDefinitions.R
\name{createCohortSubsetDefinition}
\alias{createCohortSubsetDefinition}
\title{Create Subset Definition}
\usage{
createCohortSubsetDefinition(name, definitionId, targetOutputPairs, subsets)
}
\arguments{
\item{name}{Name of definition}

\item{definitionId}{Definition identifier}

\item{targetOutputPairs}{Vector of pairs targetCohortId, outcomeCohortId}

\item{subsets}{vector of subset instances to apply}
}
\description{
Create subset definition from subset objects
}
