% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createCohortSubsetDefinition}
\alias{createCohortSubsetDefinition}
\title{Create Subset Definition}
\usage{
createCohortSubsetDefinition(targetCohortIds, outputCohortId, subsets)
}
\arguments{
\item{targetCohortIds}{cohort identifiers to apply subset functions to}

\item{subsets}{vector of subset instances to apply}
}
\description{
Create subset definition from subset objects
}
\examples{
mySubsets <- c(
 createCohortSubset(id = 12, name = "foo", cohortJson = cohortDefinition)
)

subsetDef <- createSubsetDefinition(c(1,3,4), 11, mySubsets)
}
