% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateCohorts.R
\name{createCohortTemplateDefintion}
\alias{createCohortTemplateDefintion}
\title{Create Cohort Template Definition}
\usage{
createCohortTemplateDefintion(
  name,
  templateRefFun,
  executeFun,
  templateRefArgs,
  executeArgs,
  requireConnectionRefs
)
}
\description{
construct a cohort template definition
}
