% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateCohorts.R
\name{createCohortTemplateDefintion}
\alias{createCohortTemplateDefintion}
\title{Create Cohort Template Definition}
\usage{
createCohortTemplateDefintion(
  name,
  templateSql,
  references,
  sqlArgs = list(),
  translateSql = TRUE
)
}
\description{
construct a cohort template definition
}
