% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createDemographicCriteria}
\alias{createDemographicCriteria}
\title{Create demographic criteria}
\usage{
createDemographicCriteria(ageMin = 0, ageMax = 9999, gender = "")
}
\arguments{
\item{ageMin}{age demographics}

\item{ageMax}{age demographics}

\item{gender}{gender demographics
# TODO: more criteria than this - calendarYearMin/Max, what else?}
}
\description{
Create demographic criteria
}
