% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createDemographicSubset}
\alias{createDemographicSubset}
\title{Create DemographicCriteria Subset}
\usage{
createDemographicSubset(id, name, ...)
}
\arguments{
\item{id}{Id number}

\item{name}{char name}

\item{...}{Demographic criteria @seealso createDemographicCriteria}
}
\description{
Create DemographicCriteria Subset
}
