% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createDemographicSubset}
\alias{createDemographicSubset}
\title{Create createDemographicSubset Subset}
\usage{
createDemographicSubset(
  id,
  name,
  ageMin = 0,
  ageMax = 9999,
  gender = NULL,
  race = NULL,
  ethnicity = NULL
)
}
\arguments{
\item{id}{Id number}

\item{name}{char name}

\item{ageMin}{The minimum age}

\item{ageMax}{The maximum age}

\item{gender}{Gender demographics - concept ID list}

\item{race}{Race demographics - concept ID list}

\item{ethnicity}{Ethnicity demographics - concept ID list}
}
\description{
Create createDemographicSubset Subset
}
