% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createDemographicSubset}
\alias{createDemographicSubset}
\title{Create createDemographicSubset Subset}
\usage{
createDemographicSubset(
  name = NULL,
  ageMin = 0,
  ageMax = 99999,
  gender = NULL,
  race = NULL,
  ethnicity = NULL
)
}
\arguments{
\item{name}{Optional char name}

\item{ageMin}{The minimum age}

\item{ageMax}{The maximum age}

\item{gender}{Gender demographics - concepts - 0, 8532, 8507, 0 "male", "female".
Any string that is not (case insensitive) "male" or "female" is converted to gender concept 0
https://www.ohdsi.org/web/wiki/doku.php?id=documentation:vocabulary:gender
Specific concept ids not in this set can be used but are not explicitly validated}

\item{race}{Race demographics - concept ID list}

\item{ethnicity}{Ethnicity demographics - concept ID list}
}
\description{
Create createDemographicSubset Subset
}
