% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createLimitSubset}
\alias{createLimitSubset}
\title{Create Observation Criteria Subset}
\usage{
createLimitSubset(id, name, priorTime, followUpTime, limitTo)
}
\arguments{
\item{priorTime}{Required prior observation window}

\item{followUpTime}{Required post observation window}

\item{limitTo}{charachter one of:
                             "earliest" - only first entry in patient history
                             "earliestRemaining" - only first entry after washout set by priorTime
                             "last" - only last entry in patient history
                             "lastRemaining" - only last entry in patient history inside

                         Note, when using earliest and last with follow up and washout, patients with events
                         outside this will be censored.}
}
\description{
Subset cohorts using specificed observation criteria
}
