% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createLimitSubset}
\alias{createLimitSubset}
\title{Create Observation Criteria Subset}
\usage{
createLimitSubset(id, name, priorTime, followUpTime, limitTo)
}
\arguments{
\item{priorTime}{Required prior observation window}

\item{followUpTime}{Required post observation window}

\item{limitTo}{character one of:
                             "firstEver" - only first entry in patient history
                             "earliestRemaining" - only first entry after washout set by priorTime
                             "latestRemaining" -  the latest remaining after washout set by followUpTime
                             "lastEver" - only last entry in patient history inside

                         Note, when using firstEver and lastEver with follow up and washout, patients with events
                         outside this will be censored. The "firstEver" and "lastEver" are applied first. 
                         The "earliestRemaining" and "latestRemaining" are applied after all other limit 
                         criteria are applied (i.e. after applying prior/post time and calendar time).}
}
\description{
Subset cohorts using specificed observation criteria
}
