% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createLimitSubset}
\alias{createLimitSubset}
\title{Create Limit Subset Operator}
\usage{
createLimitSubset(...)
}
\arguments{
\item{...}{Arguments passed to the underlying operator.}
}
\description{
Subset cohorts using specified limit criteria.
deprecated This function is deprecated. Please use `createLimitSubsetOperator()` instead.
}
