% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateImplementations.R
\name{createSnomedCohortTemplateDefinition}
\alias{createSnomedCohortTemplateDefinition}
\title{Create SNOMED cohort Template Definition}
\usage{
createSnomedCohortTemplateDefinition(
  indentifierExpression = "concept_id * 1000",
  cdmDatabaseSchema,
  conditionsTable = "cohort_conditions_ref",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema,
  priorObservationPeriod = 365,
  vocabularyDatabaseSchema = cdmDatabaseSchema
)
}
\arguments{
\item{indentifierExpression}{an expression for setting the cohort id for the resulting cohort. Must produce unique ids}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{conditionsTable}{reference table to store condition cohorts}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{priorObservationPeriod}{(optional) required prior observation period for individuals}
}
\value{
a CohortTemplateDefinition instance
}
\description{
Template cohort definition for all OHDSI standard conditions
This cohort will use the vocaublary tables to automaticall generate a set of cohorts that have the
cohortId = conceptId * 1000 + 4, note that this can be customised with the "identifierExpression" if you are using this
with other cohorts you may wish to change this to allow uniqueness
}
