% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateImplementations.R
\name{createSnomedCohortTemplateDefinition}
\alias{createSnomedCohortTemplateDefinition}
\title{Create SNOMED Cohort Template Definition}
\usage{
createSnomedCohortTemplateDefinition(
  connection,
  identifierExpression = "concept_id * 1000",
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema,
  priorObservationPeriod = 365,
  requireSecondDiagnosis = FALSE,
  nameSuffix = "",
  vocabularyDatabaseSchema = cdmDatabaseSchema
)
}
\arguments{
\item{connection}{Database connection object}

\item{identifierExpression}{An expression for setting the cohort id for the resulting cohort. Must produce unique ids}

\item{cdmDatabaseSchema}{CDM database schema}

\item{tempEmulationSchema}{Temporary emulation schema}

\item{cohortDatabaseSchema}{Cohort database schema}

\item{priorObservationPeriod}{(optional) Required prior observation period for individuals}

\item{vocabularyDatabaseSchema}{Vocabulary database schema}

\item{conditionsTable}{Reference table to store condition cohorts}
}
\value{
A CohortTemplateDefinition instance
}
\description{
Template cohort definition for all OHDSI standard conditions. The cohortId = conceptId * 1000.
The "identifierExpression" can be customized for uniqueness.
}
