% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{createSubsetCohortWindow}
\alias{createSubsetCohortWindow}
\title{A definition of subset functions to be applied to a set of cohorts}
\usage{
createSubsetCohortWindow(startDay, endDay, targetAnchor)
}
\arguments{
\item{startDay}{The start day for the window}

\item{endDay}{The end day for the window}

\item{targetAnchor}{To anchor using the target cohort's start date or end date}
}
\value{
a SubsetCohortWindow instance
}
\description{
A definition of subset functions to be applied to a set of cohorts
}
