% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateUnion.R
\name{createUnionCohortTemplate}
\alias{createUnionCohortTemplate}
\title{Create cohort template to union multiple cohorts}
\usage{
createUnionCohortTemplate(
  cohortDatabaseSchema,
  cohortTable,
  cohortIds,
  unionCohortId
)
}
\arguments{
\item{cohortDatabaseSchema}{Schema where cohort tables are stored.}

\item{cohortTable}{Name of the cohort table.}

\item{cohortIds}{A vector of `cohort_definition_id` values for the input cohorts.}

\item{unionCohortId}{The `cohort_definition_id` for the resulting union cohort.}
}
\description{
Create cohort template to union multiple cohorts
}
