% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{DemographicCriteria}
\alias{DemographicCriteria}
\title{to List}
\description{
Representation of demographic settings to be used in a subset instance
}
\details{
Demographics settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ageMin}}{Int between 0 and 9999 - minimum age}

\item{\code{ageMax}}{Int between 0 and 9999 - maximum age}

\item{\code{gender}}{character string denoting gender}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-toList}{\code{DemographicCriteria$toList()}}
\item \href{#method-toJSON}{\code{DemographicCriteria$toJSON()}}
\item \href{#method-isEqualTo}{\code{DemographicCriteria$isEqualTo()}}
\item \href{#method-clone}{\code{DemographicCriteria$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCriteria$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCriteria$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare Subset to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCriteria$isEqualTo(criteria)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{criteria}}{DemographicCriteria instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCriteria$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
