% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{DemographicSubsetOperator}
\alias{DemographicSubsetOperator}
\title{Public Fields}
\description{
Public Fields

Public Fields

to List

is Equal to
}
\details{
Criteria Subset
}
\section{Super class}{
\code{\link[CohortGenerator:SubsetOperator]{CohortGenerator::SubsetOperator}} -> \code{DemographicSubsetOperator}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{criteria}}{DemographicCriteria to subset to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-publicFields}{\code{DemographicSubsetOperator$publicFields()}}
\item \href{#method-toList}{\code{DemographicSubsetOperator$toList()}}
\item \href{#method-isEqualTo}{\code{DemographicSubsetOperator$isEqualTo()}}
\item \href{#method-clone}{\code{DemographicSubsetOperator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="classname">}\href{../../CohortGenerator/html/SubsetOperator.html#method-classname}{\code{CohortGenerator::SubsetOperator$classname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="initialize">}\href{../../CohortGenerator/html/SubsetOperator.html#method-initialize}{\code{CohortGenerator::SubsetOperator$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CohortGenerator" data-topic="SubsetOperator" data-id="toJSON">}\href{../../CohortGenerator/html/SubsetOperator.html#method-toJSON}{\code{CohortGenerator::SubsetOperator$toJSON()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-publicFields"></a>}}
\if{latex}{\out{\hypertarget{method-publicFields}{}}}
\subsection{Method \code{publicFields()}}{
Publicly settable fields of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicSubsetOperator$publicFields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicSubsetOperator$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare Subset to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicSubsetOperator$isEqualTo(subsetOperatorB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subsetOperatorB}}{A subset to test equivalence to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicSubsetOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
