% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSave.R
\name{exportCohortStatsTables}
\alias{exportCohortStatsTables}
\title{Export the cohort statistics tables to the file system}
\usage{
exportCohortStatsTables(
  connectionDetails,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTableNames = getCohortTableNames(),
  cohortStatisticsFolder,
  incremental = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{The schema to hold the cohort tables. Note that for
SQL Server, this should include both the database and schema
name, for example 'scratch.dbo'.}

\item{cohortTableNames}{The names of the cohort statistics tables. See \code{\link{getCohortTableNames}}
for more details.}

\item{cohortStatisticsFolder}{The path to the folder where the cohort statistics folder 
where the results will be written}

\item{incremental}{If \code{incremental = TRUE}, results are written to update values instead of 
overwriting an existing results}
}
\description{
This function retrieves the data from the cohort statistics tables and
writes them to the inclusion statistics folder specified in the function
call.
}
