% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemplateCohorts.R
\name{getTemplateDefinitions}
\alias{getTemplateDefinitions}
\title{Extract template definitions from a cohort definition set}
\usage{
getTemplateDefinitions(cohortDefinitionSet)
}
\arguments{
\item{cohortDefinitionSet}{The \code{cohortDefinitionSet} argument must be a data frame with 
the following columns: \describe{
\item{cohortId}{The unique integer identifier of the cohort} 
\item{cohortName}{The cohort's name}
\item{sql}{The OHDSI-SQL used to generate the cohort}}
Optionally, this data frame may contain: \describe{
\item{json}{The Circe JSON representation of the cohort}}}
}
\description{
Extract template definitions from a cohort definition set
}
\seealso{
Other templateCohorts: 
\code{\link{CohortTemplateDefinition}},
\code{\link{addCohortTemplateDefintion}()},
\code{\link{addSqlCohortDefinition}()},
\code{\link{createCohortTemplateDefintion}()}
}
\concept{templateCohorts}
