% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{isFormattedForDatabaseUpload}
\alias{isFormattedForDatabaseUpload}
\title{Is the data.frame formatted for uploading to a database?}
\usage{
isFormattedForDatabaseUpload(x, warn = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{warn}{When TRUE, display a warning of any columns are not in snake
case format}
}
\value{
Returns TRUE if all columns are snake case format. If warn == TRUE,
the function will emit a warning on the column names that are not in snake
case format.
}
\description{
This function is used to check a data.frame to ensure all
column names are in snake case format.
}
