% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{isSnakeCase}
\alias{isSnakeCase}
\title{Used to check if a string is in snake case}
\usage{
isSnakeCase(x)
}
\arguments{
\item{x}{The string to evaluate}
}
\value{
TRUE if the string is in snake case
}
\description{
This function is used check if a string conforms to
the snake case format.
}
