% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{loadRenderTranslateSql}
\alias{loadRenderTranslateSql}
\title{Load, render, and translate a SQL file in this package.
This is done in place of using SqlRender::loadRenderTranslateSql
otherwise unit tests will not function properly.}
\usage{
loadRenderTranslateSql(
  sqlFilename,
  dbms = "sql server",
  ...,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  warnOnMissingParameters = TRUE
)
}
\description{
NOTE: This function does not support dialect-specific SQL translation
at this time.
}
