% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omopCdmDrugExposure}
\alias{omopCdmDrugExposure}
\title{OMOP CDM Drug Exposure Sample Data}
\format{
A data frame with 8 rows and 5 variables:
\describe{
  \item{drug_exposure_id}{A unique identifier for the drug exposure}
  \item{person_id}{An integer representing the patient}
  \item{drug_concept_id}{An integer concept ID representing the drug concept}
  \item{drug_exposure_start_date}{Drug start date}
  \item{drug_exposure_end_date}{Drug end date}
}
}
\source{
Fictional data for demonstration.
}
\usage{
omopCdmDrugExposure
}
\description{
A data set containing sample drug exposures for 2 drugs
}
\keyword{datasets}
