% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{readCsv}
\alias{readCsv}
\title{Used to read a csv file}
\usage{
readCsv(file)
}
\arguments{
\item{file}{The csv file to read.}
}
\value{
A tibble with the CSV contents
}
\description{
This function is used to centralize the function for reading
CSV files across the HADES ecosystem.
}
