% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetDefinitions.R
\name{saveCohortSubsetDefinition}
\alias{saveCohortSubsetDefinition}
\title{Save cohort subset definitions to json}
\usage{
saveCohortSubsetDefinition(
  subsetDefinition,
  subsetJsonFolder = "inst/cohort_subset_definitions/"
)
}
\arguments{
\item{subsetDefinition}{The subset definition object {@seealso CohortSubsetDefinition}}

\item{subsetJsonFolder}{Defines the folder to store the subset JSON}
}
\description{
This is generally used as part of saveCohortDefinitionSet
}
