% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsets.R
\name{SubsetCohortWindow}
\alias{SubsetCohortWindow}
\title{to List}
\description{
Representation of a time window to use when subsetting a target cohort with a subset cohort
}
\details{
SubsetCohortWindow settings
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{startDay}}{Integer}

\item{\code{endDay}}{Integer}

\item{\code{targetAnchor}}{Boolean}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-toList}{\code{SubsetCohortWindow$toList()}}
\item \href{#method-toJSON}{\code{SubsetCohortWindow$toJSON()}}
\item \href{#method-isEqualTo}{\code{SubsetCohortWindow$isEqualTo()}}
\item \href{#method-clone}{\code{SubsetCohortWindow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
List representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubsetCohortWindow$toList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
json serialized representation of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubsetCohortWindow$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEqualTo"></a>}}
\if{latex}{\out{\hypertarget{method-isEqualTo}{}}}
\subsection{Method \code{isEqualTo()}}{
Compare SubsetCohortWindow to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubsetCohortWindow$isEqualTo(criteria)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{criteria}}{SubsetCohortWindow instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubsetCohortWindow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
