% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CsvHelper.R
\name{writeCsv}
\alias{writeCsv}
\title{Used to write a csv file}
\usage{
writeCsv(x, file, warnOnUploadRuleViolations = TRUE)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{The csv file to write.}

\item{warnOnUploadRuleViolations}{When TRUE, this function will provide
warning messages that may indicate if the data is stored in a format in the
CSV that may cause problems when uploading to a database.}
}
\value{
Returns the input x invisibly.
}
\description{
This function is used to centralize the function for writing
CSV files across the HADES ecosystem. This function may also raise warnings
if the data is stored in a format that will not work with the HADES standard
for uploading to a results database.
}
